refactor(parse): migrate acpi, battery, proc, and control flow variables to registry
  Move ~70 more text objects from the core.cc if-else chain into
  distributed CONKY_REGISTER_VARIABLES blocks in their feature files.

  Migrated groups:
  - ACPI/battery (common.cc, #ifndef __OpenBSD__): acpitemp,
    acpiacadapter, acpifan, battery, battery_short, battery_status,
    battery_time, battery_percent, battery_power_draw, battery_bar
  - Freq: freq, freq_g
  - Control flow: if_empty, if_existing, if_running (all platform
    variants), if_updatenr, no_update, eval, blink, include
  - Misc: loadgraph, github_notifications, stock
  - Proc (proc.cc): all 40 pid_* variables, cmdline_to_pid

  Infrastructure additions:
  - arg_object_variable: template helper for variables with live
    sub-expression arguments (pid_* pattern), inlines
    extract_object_args_to_sub
  - print_cb type alias for print callback signature
  - CONKY_CONCAT macro in macros.h for unique static init names,
    enabling multiple CONKY_REGISTER_VARIABLES per translation unit

  313 lines removed from core.cc, ~26% of the chain cleaned up.
