+++ src/macros.h
@@ -68,4 +68,8 @@
   } while (0)
 #endif /* NDEBUG */
 
+// Token concatenation with macro expansion.
+#define CONKY_CONCAT_(a, b) a##b
+#define CONKY_CONCAT(a, b) CONKY_CONCAT_(a, b)
+
 #endif /* _CONKY_MACROS_H_ */
